requirement.txt
torchvision
torchio
medpy
scikit-learn
tensorboard
monai
albumentations